src/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
pydantic-settings
python-jose
bcrypt==4.0.1
passlib[bcrypt]
python-dateutil